+++ collective/behavior/relatedmedia/interfaces.py
@@ -47,6 +47,13 @@ class IRelatedMediaSettings(Interface):
         required=False,
     )
 
+    update_leadimage = schema.Bool(
+        title=_(u'Set first related image as leadimage?'),
+        description=_(u'This is applied on any change.'),
+        default=False,
+        required=False,
+    )
+
     open_attachment_in_new_window = schema.Bool(
         title=_(u'Open Attachment links in new window'),
         default=True,